+++ lib/quick_travel/constants.rb
@@ -1,5 +1,4 @@
 ### DELETE ME!  Left here only so you can pull out the constants for the few that we should be keeping.
 
 module QuickTravel
-  LOGIN_URL  = '/parties/login' # CMS_SSL_URL + '/parties/login'
 end