+++ lib/quick_travel/party.rb
@@ -1,6 +1,6 @@
 module QuickTravel
   class Party < Adapter
-    LOGIN_URL = '/login.json'
+    LOGIN_URL = '/api/login.json'
 
     attr_reader :phone, :mobile, :email   # if has a contact
     attr_reader :post_code, :country_id   # if has an address