+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## [2.9.1]
+ - Allow zero pricing for extra pick items
+
 ## [2.9.0]
 ### Changed
 - Use new API login